requirements.txt
dspy-ai==2.5.43
fastapi==0.115.6
numpy==2.2.0
pillow==11.0.0
pytest-playwright==0.6.2
torch==2.5.1
uvicorn==0.34.0